判断工程是否已经存在了

+++ dev_cpp_vnc_eclipse/scripts/vnc_startup.sh
@@ -29,7 +29,10 @@ ln -s /usr/bin/firefox /root/Desktop/firefox
 ln -s /usr/bin/evince /root/Desktop/evince
 ln -s /usr/bin/goldendict /root/Desktop/goldendict
 mkdir -P /root/work_dir/
-tar -zxf /root/eclipse_workspace.tar.gz -C /root/work_dir/ && rm -f /root/eclipse_workspace.tar.gz
+if [ ! -d "/root/work_dir/eclipse_workspace/" ] ; then
+        tar -zxf /root/eclipse_workspace.tar.gz -C /root/work_dir/ && rm -f /root/eclipse_workspace.tar.gz
+fi
+
 
 # log connect options
 echo -e "\n------------------ VNC environment started ------------------"
